Test custom local repositories passed through launcher params with Scala 3.5.0-RC1-fakeversion-bin-SNAPSHOT
