+++ modules/integration/src/test/scala/scala/cli/integration/BspTestDefinitions.scala
@@ -2107,72 +2107,6 @@ abstract class BspTestDefinitions extends ScalaCliSuite with TestScalaVersionArg
     }
   }
 
-  test("BSP respects --cli-default-scala-version & --predefined-repository launcher options") {
-    val sv = "3.5.0-RC1"
-    val inputs = TestInputs(
-      os.rel / "simple.sc" -> s"""assert(dotty.tools.dotc.config.Properties.simpleVersionString == "$sv")"""
-    )
-    inputs.fromRoot { root =>
-      os.proc(TestUtil.cli, "bloop", "exit", "--power").call(cwd = root)
-      val localRepoPath = root / "local-repo"
-      val artifactNames = Seq(
-        "scala3-compiler_3",
-        "scala3-staging_3",
-        "scala3-tasty-inspector_3",
-        "scala3-sbt-bridge"
-      )
-      for { artifactName <- artifactNames } {
-        val csRes = os.proc(
-          TestUtil.cs,
-          "fetch",
-          "--cache",
-          localRepoPath,
-          s"org.scala-lang:$artifactName:$sv"
-        )
-          .call(cwd = root)
-        expect(csRes.exitCode == 0)
-      }
-      os.proc(
-        TestUtil.cli,
-        "--cli-default-scala-version",
-        sv,
-        "--predefined-repository",
-        (localRepoPath / "https" / "repo1.maven.org" / "maven2").toNIO.toUri.toASCIIString,
-        "setup-ide",
-        "simple.sc",
-        "--with-compiler",
-        "--offline",
-        "--power"
-      )
-        .call(cwd = root)
-      val ideOptionsPath = root / Constants.workspaceDirName / "ide-options-v2.json"
-      expect(ideOptionsPath.toNIO.toFile.exists())
-      val ideEnvsPath = root / Constants.workspaceDirName / "ide-envs.json"
-      expect(ideEnvsPath.toNIO.toFile.exists())
-      val ideLauncherOptionsPath = root / Constants.workspaceDirName / "ide-launcher-options.json"
-      expect(ideLauncherOptionsPath.toNIO.toFile.exists())
-      val jsonOptions     = List("--json-options", ideOptionsPath.toString)
-      val launcherOptions = List("--json-launcher-options", ideLauncherOptionsPath.toString)
-      val envOptions      = List("--envs-file", ideEnvsPath.toString)
-      val bspOptions      = jsonOptions ++ launcherOptions ++ envOptions
-      withBsp(inputs, Seq("."), bspOptions = bspOptions, reuseRoot = Some(root)) {
-        (_, _, remoteServer) =>
-          async {
-            val targets = await(remoteServer.workspaceBuildTargets().asScala)
-              .getTargets.asScala
-              .filter(!_.getId.getUri.contains("-test"))
-              .map(_.getId())
-            val compileResult =
-              await(remoteServer.buildTargetCompile(new b.CompileParams(targets.asJava)).asScala)
-            expect(compileResult.getStatusCode == b.StatusCode.OK)
-            val runResult =
-              await(remoteServer.buildTargetRun(new b.RunParams(targets.head)).asScala)
-            expect(runResult.getStatusCode == b.StatusCode.OK)
-          }
-      }
-    }
-  }
-
   private def checkIfBloopProjectIsInitialised(
     root: os.Path,
     buildTargetsResp: b.WorkspaceBuildTargetsResult
